+++ packages/sdk-sui/src/index.ts
@@ -1,4 +1,5 @@
 export * from './const';
+export * from './utils/createSuiClient';
 export * from './web3Sdk/claimLBTC';
 export * from './web3Sdk/getBalance';
 export * from './web3Sdk/getBasculeDepositStatus';